feat(core/raw/buf): Reduce one allocation by `Arc::from_iter`
